Refactor script-templates Makefile.
This is the first step to add new scheduler/ressource manager
to magpie code base. I plan to submit a Moab/Torque patch
which would use pdsh or mpiexec instead of srun. This also
why the only template target currently ends with the suffix
".srun", I plan to add .pdsh and .mpi.