Added per job support for hadoop hdfs and zookeeper data dir
This patch allows a user to run per job instances of hdfs and the
zookeeper data dir. When marked 'yes' it will append the job id
to the directory isolating the data. This allows for running
multiple jobs using different data sets in parallel without
stepping on each other. The data does not get automatically
deleted after a job.