Re-organize README so patching instructions are located along with the project section
Re-organize README so advanced usage instructions are located along with the project section

Re-organize README so testing information are located along with the project section

Re-organize README so environment variable exports are located along with the project section.

Re-organize README so convenience script information is located along with the project section.