Remove 'classifywikipedia' example job from Mahout
Example downloads "latest" files from wikipedia, which can change
and may not be up to date.  So sometimes Mahout can file if the
files become out of date and output could differ wildly.  So it's not
a good example.  Leave clustersyntheticcontrol as primary example
since it is fixed on a specific data source.