Experimental support for Flux
This was predominantly an exercise and to find subtle issues within
Flux when running w/ Magpie.  It was experimented on around v0.5.0
release of Flux.  See following issues as examples
as discussion that occurred:

https://github.com/flux-framework/flux-core/issues/883
https://github.com/flux-framework/flux-core/issues/864
https://github.com/flux-framework/flux-core/issues/866

For info on how to run see magpie.sbatch-srun-spark-test.sh example.
See information at the top for how to run and see info at the bottom
for workarounds for flux.